+++ community/src/main/java/community/cmm/service/impl/FileServiceImpl.java
@@ -31,15 +31,46 @@ public class FileServiceImpl implements FileService {
 	/** FileDAO DI */
 	@Autowired
 	FileDAO fileDAO;
+	
+	
+	/** 업로드 파일 DB 등록 */
+	@Override
+	public Integer insertFile(FileVO vo) throws Exception {
+		// 파일 순번 생성 로직
+		Integer checkFileId = fileDAO.selectLastFileSn(vo);
+		Integer nextIdNum = 1;
+		if (checkFileId != null) {
+			nextIdNum += checkFileId;
+		}
+		vo.setFileSn(nextIdNum.toString());
+		fileDAO.insertFile(vo);
+
+		return nextIdNum;
+	}
+	
+	
+	/** 게시글 내용으로부터 업로드된 이미지의 아이디 추출 */
+	@Override
+	public List<String> extractImageIds(String inputString) throws Exception {
+		List<String> resultList = new ArrayList<>();
+		
+		Pattern pattern = Pattern.compile("data-image-id=\"([^\"]+)\"");
+		Matcher matcher = pattern.matcher(inputString);
+		while (matcher.find()) {
+			resultList.add(matcher.group(1));
+		}
+
+		return resultList;
+	}
 
 	
-	/** 업로드 파일 상태정보 수정 */
+	/** 업로드 파일 DB정보 갱신 */
 	@Override
 	public String updateFileUseAt(FileVO vo) throws Exception {
 		List<String> extractImageIds = extractImageIds(vo.getBoardCn()); // 작성 또는 수정 게시글 내용에 포함된 이미지 정보 리스트
 		int updateUseAt = 0;
 		
-		// 게시글에 포함된 이미지 정보를 db에 최종등록 및 썸네일 생성
+		// 게시글에 포함된 이미지 상태정보를 DB에 최종 갱신 및 썸네일 생성
 		boolean isFirst = true;
 		for (String imageId : extractImageIds) {
 			String[] imageIdParts = imageId.split("_(?=[^_]+$)");
@@ -47,30 +78,30 @@ public class FileServiceImpl implements FileService {
 				log.info("게시글 내용으로부터 아이디 추출 및 분리 성공");
 				vo.setAtchFileId(imageIdParts[0]);
 				vo.setFileSn(imageIdParts[1]);
-				if (!"N".equals(vo.getUseAtBoard())) { // 게시글삭제인경우 실행 x
+				
+				// 임시 상태의 이미지를 사용으로 갱신
+				if (!"N".equals(vo.getUseAtBoard())) { // 게시글이 삭제되지 않은 경우
 					vo.setUseAt("Y");
 					updateUseAt += fileDAO.updateUseAt(vo);
+					
+					// 썸네일 생성 작업
 					if (isFirst) { // 글내용의 첫 이미지인 경우에만 동작
 						log.info("썸네일 작업 시작");
 						log.info("썸네일 작업을 위한 vo 정보", vo.toString());
-						thumbnail(vo); // 썸네일 시작
+						FileVO thumbnail = thumbnail(vo); // 썸네일 생성
+						updateThumbnail(thumbnail); // 썸네일 정보 등록 or 갱신
 					}
 				}
 				isFirst = false;
 			}
 		}
 		
-		if (isFirst || "N".equals(vo.getUseAtBoard())) { // 게시글 내용에 이미지가 없을 때 동작
+		// 썸네일 삭제
+		if (isFirst || "N".equals(vo.getUseAtBoard())) { // 게시글 내용에 이미지가 없거나 게시글이 삭제된 경우
 			log.info("썸네일 제거 체크 : {}", vo.getAtchFileId());
-			vo.setFileSn("0");
-			if ("N".equals(vo.getUseAtBoard())) {
-				vo.setUseAt("Y");
-			} else {
-				vo.setUseAt("N");
-			}
-			
 			if (StringUtils.hasText(vo.getAtchFileId())) {
-				FileVO resultFile = fileDAO.selectFile(vo);
+				vo.setFileSn("0");
+				FileVO resultFile = fileDAO.selectFile(vo); // 썸네일 정보 불러오기
 				if (resultFile != null) {
 					log.info("썸네일 제거 시작");
 					File file = new File(resultFile.getFileStreCours() + resultFile.getStreFileNm()+ "." + resultFile.getFileExtsn());
@@ -84,11 +115,11 @@ public class FileServiceImpl implements FileService {
 			}
 		}
 		
-		vo.setUseAt("T");
+		vo.setUseAt("!N");
 		vo.setFileSn("");
 		List<FileVO> resultFileList = fileDAO.selectFileList(vo); // db에 등록된 이미지 정보 리스트
 		
-		// db에 등록된 이미지정보와 실제 게시글에 포함된 이미지 정보 비교 후 실제 게시글에 이미지가 없으면 db 데이터 상태 변경
+		// db에 등록된 이미지정보와 실제 게시글에 포함된 이미지 정보 비교 후 실제 게시글에 이미지가 없으면 이미지 상태 변경
 		for (FileVO file : resultFileList) {
 			String resultImageId = file.getAtchFileId() + "_" + file.getFileSn();
 			log.info("게시글에 포함되었는지 체크할 db 데이터 : {}", resultImageId);
@@ -103,7 +134,11 @@ public class FileServiceImpl implements FileService {
 		return vo.getAtchFileId();
 	}
 	
+	
+	
+	
 	/** 썸네일 작업 */
+	@Override
 	public FileVO thumbnail(FileVO vo) throws Exception {
 		log.info("썸네일 작업 시작");
 		
@@ -111,7 +146,6 @@ public class FileServiceImpl implements FileService {
 		FileVO preFile = fileDAO.selectFile(vo);
 		
 		log.info("썸네일 원본 이미지 정보 가져오기 완료");
-		
 
 		// 썸네일 생성할 폴더, 파일명 세팅
 		vo.setFileStreCours(preFile.getFileStreCours() + "thumbnail" + File.separator);
@@ -157,66 +191,47 @@ public class FileServiceImpl implements FileService {
 		String fileExtension = fileName.substring(fileName.lastIndexOf(".") + 1);
 		vo.setFileExtsn(fileExtension);
 		
-		log.info("썸네일 정보 가져오기 완료");
 		
 		// 기존 썸네일 존재여부 체크
 		vo.setFileSn("0");
 		vo.setUseAt("any");
-		log.info("썸네일 정보등록을 위한 vo 정보", vo.toString());
+		log.info("썸네일 정보등록을 위한 vo 정보 {}", vo.toString());
 		
+		
+		return vo;
+	}
+	
+	/** 썸네일 DB 정보 갱신 */
+	@Override
+	public void updateThumbnail (FileVO vo) throws Exception {
 		FileVO resultFile = fileDAO.selectFile(vo);
 		log.info("썸네일 정보 셋 완료");
-		if (resultFile != null) {
-			log.info("썸네일 정보 갱신");
-			File oldThumnail = new File(resultFile.getFileStreCours() + resultFile.getStreFileNm()+ "." + resultFile.getFileExtsn());
-			if (oldThumnail.exists()) {
-				oldThumnail.delete();
-				log.info("삭제 파일 존재 확인 : {}", resultFile.getFileStreCours() + resultFile.getStreFileNm());
+		vo.setUseAt("T");
+		// 썸네일 변경(기존썸네일 디스크에서 제거) or 등록 
+		if (resultFile != null && StringUtils.hasText(vo.getStreFileNm())) {
+			if (!resultFile.getStreFileNm().equals(vo.getStreFileNm())) { // 기존 썸네일과 새로 생성된 썸네일이 같지 않으면 기존 썸네일 삭제
+				log.info("썸네일 정보 갱신");
+				File oldThumnail = new File(resultFile.getFileStreCours() + resultFile.getStreFileNm()+ "." + resultFile.getFileExtsn());
+				// 기존썸네일 삭제
+				if (oldThumnail.exists()) {
+					oldThumnail.delete();
+					log.info("기존 썸네일 삭제 완료 : {}", resultFile.getFileStreCours() + resultFile.getStreFileNm());
+				}
+				// 새로운 썸네일 DB 정보 갱신
+				fileDAO.updateFile(vo);
+				log.info("썸네일 정보 갱신 완료");
 			}
-			vo.setUseAt("Y");
-			fileDAO.updateFile(vo);
-		} else {
-			log.info("썸네일 정보 등록");
-			vo.setUseAt("Y");
+		} else { // 신규로 썸네일 DB 정보 등록
 			fileDAO.insertFile(vo);
+			log.info("썸네일 정보 등록 완료");
 		}
 		
-		
-		
-		log.info("썸네일 useAt 갱신");
+		// 썸네일 상태 정보 갱신
+		vo.setUseAt("Y");
 		fileDAO.updateUseAt(vo); //썸네일 useAt 갱신
-		return null;
+		log.info("썸네일 상태 정보 갱신 완료 {}", vo.getUseAt());
 	}
 	
-	/** 게시글 내용으로부터 업로드된 이미지의 아이디 추출 */
-	@Override
-	public List<String> extractImageIds(String inputString) throws Exception {
-		List<String> resultList = new ArrayList<>();
-		Pattern pattern = Pattern.compile("data-image-id=\"([^\"]+)\"");
-		Matcher matcher = pattern.matcher(inputString);
-
-		while (matcher.find()) {
-			resultList.add(matcher.group(1));
-		}
-
-		return resultList;
-	}
-	
-
-	/** 업로드 파일 DB 등록 */
-	@Override
-	public Integer insertFile(FileVO vo) throws Exception {
-		// 파일 아이디 생성 로직
-		Integer checkFileId = fileDAO.selectLastFileSn(vo);
-		Integer nextIdNum = 1;
-		if (checkFileId != null) {
-			nextIdNum += checkFileId;
-		}
-		vo.setFileSn(nextIdNum.toString());
-		fileDAO.insertFile(vo);
-
-		return nextIdNum;
-	}
 
 	/** 스마트 에디터 이미지 업로드 */
 	@Override
@@ -253,7 +268,9 @@ public class FileServiceImpl implements FileService {
 			while ((num = is.read(buffer)) != -1)
 				os.write(buffer, 0, num);
 		}
-		vo.setUseAt("T"); // 임시저장
+		
+		// DB 파일정보 삽입(임시파일)
+		vo.setUseAt("T");
 		Integer fileSn = insertFile(vo);
 
 		// 스마트 에디터에 전송할 파일정보
